+++ frontend/client/viewer/components/data-table-row-details.tsx
@@ -1,34 +1,13 @@
 import * as React from 'react';
 import Table from '@mui/material/Table';
-import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
-import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import { Button, Paper } from '@mui/material';
-import { ClientTaskLocalStorageData, ClientTaskTypeDescriptors, PocketData, ServerTaskLocalStorageData, ServerTaskType, ServerTaskTypeDescriptors } from "../../custom-types";
-import { downloadDockingResult } from '../../tasks/server-docking-task';
-import { ClientTaskType } from "../../custom-types";
+import { PocketData } from "../../custom-types";
+import { TasksTable } from "./tasks-table";
 
 export default function DataTableRowDetails(props: { pocket: PocketData; setTab: (tab: number, initialPocket?: number) => void, structureId: string; }) {
     const pocket = props.pocket;
-
-    let serverTasks = localStorage.getItem(`${props.structureId}_serverTasks`);
-    if (!serverTasks) serverTasks = "[]";
-    const serverTasksParsed: ServerTaskLocalStorageData[] = JSON.parse(serverTasks);
-
-    let clientTasks = localStorage.getItem(`${props.structureId}_clientTasks`);
-    if (!clientTasks) clientTasks = "[]";
-    const clientTasksParsed: ClientTaskLocalStorageData[] = JSON.parse(clientTasks);
-
-    const handleResultClick = (serverTask: ServerTaskLocalStorageData) => {
-        switch (serverTask.type) {
-            case ServerTaskType.Docking:
-                downloadDockingResult(serverTask.params[0], serverTask.responseData[0].url, pocket.rank);
-                break;
-            default:
-                break;
-        }
-    };
 
     const handleCreateTask = () => {
         //the tab index is 2
@@ -84,45 +63,7 @@ export default function DataTableRowDetails(props: { pocket: PocketData; setTab:
             &nbsp;
             <Paper>
                 {/*<span style={{ padding: 10, paddingTop: 15, fontSize: "1.25rem", display: "block" }}>Tasks for pocket {pocket.rank}</span>*/}
-                <Table size="small">
-                    <TableHead>
-                        <TableRow style={{ background: "white" }}>
-                            <TableCell style={{ fontWeight: "bold" }}>Task type</TableCell>
-                            <TableCell style={{ fontWeight: "bold" }}>Name</TableCell>
-                            <TableCell style={{ fontWeight: "bold" }}>Timestamp</TableCell>
-                            <TableCell style={{ fontWeight: "bold" }}>Status/result</TableCell>
-                        </TableRow>
-                    </TableHead>
-                    <TableBody>
-                        {clientTasksParsed.map((row: ClientTaskLocalStorageData, i) => {
-                            if (row.pocket === Number(pocket.rank)) {
-                                return (
-                                    <TableRow key={i + "_client"}>
-                                        <TableCell>{ClientTaskTypeDescriptors[row.type]}</TableCell>
-                                        <TableCell>{"-"}</TableCell>
-                                        <TableCell>{row.created}</TableCell>
-                                        <TableCell>
-                                            {(!isNaN(row.data)) ? row.data.toFixed(1) : row.data}
-                                            {row.type === ClientTaskType.Volume && " Å³"}
-                                        </TableCell>
-                                    </TableRow>
-                                );
-                            }
-                        })}
-                        {serverTasksParsed.map((row: ServerTaskLocalStorageData, i) => {
-                            if (row.pocket === Number(pocket.rank)) {
-                                return (
-                                    <TableRow key={i + "_server"}>
-                                        <TableCell>{ServerTaskTypeDescriptors[row.type]}</TableCell>
-                                        <TableCell>{row.name}</TableCell>
-                                        <TableCell>{row.created}</TableCell>
-                                        <TableCell>{row.status === "successful" ? <span onClick={() => handleResultClick(row)} style={{ color: "blue", textDecoration: "underline", cursor: "pointer" }}>successful</span> : row.status}</TableCell>
-                                    </TableRow>
-                                );
-                            }
-                        })}
-                    </TableBody>
-                </Table>
+                <TasksTable pocket={props.pocket} structureId={props.structureId} />
                 <Button variant="outlined" onClick={handleCreateTask} style={{ margin: 10 }}>Create task</Button>
             </Paper>
         </div>
